insightscribe-agent/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymongo
openai
sse_starlette
python-multipart # Added for potential file uploads, common in FastAPI setups
